requirements.txt
# ============================================================
#  ASL Alphabet Classification — Dependencies
# ============================================================

# Core DL
tensorflow>=2.13.0
numpy>=1.24.0

# Data & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# FastAPI Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
